separating composite phase space models into separate modules
