+++ halotools/empirical_models/phase_space_models/velocity_models/tests/__init__.py
@@ -1,0 +1,1 @@
+from test_nfw_isotropic_jeans import *
